feature: added view modal

+++ app/services/todo-list.tsx
@@ -1,21 +1,29 @@
 interface Body {
-  title: string,
+  title: string;
   description: string;
-  completed: boolean
+  completed: boolean;
 }
 
 const getAllTask = async () => {
   try {
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
     const response = await fetch("http://localhost:3000/tasks", {
       method: "get",
       headers: {
         "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
+        Authorization: `Bearer ${String(token)}`,
       },
     });
 
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
+
     const result = await response.json();
-    return result
+    return result;
   } catch (error) {
     console.error("Error:", error);
   }
@@ -23,31 +31,47 @@ const getAllTask = async () => {
 
 const getTaskById = async (id: number) => {
   try {
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
     const response = await fetch(`http://localhost:3000/tasks/${id}`, {
       method: "get",
       headers: {
         "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
+        Authorization: `Bearer ${String(token)}`,
       },
     });
 
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
+
     const result = await response.json();
-    return result
+    return result;
   } catch (error) {
     console.error("Error:", error);
   }
 };
 
-const createTask = async (body:Body) => {
+const createTask = async (body: Body) => {
   try {
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
     const response = await fetch("http://localhost:3000/tasks", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
+        Authorization: `Bearer ${String(token)}`,
       },
       body: JSON.stringify(body),
     });
+
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
 
     const result = await response.json();
     console.log("Success:", result);
@@ -58,14 +82,22 @@ const createTask = async (body:Body) => {
 
 const updateTask = async (body: Body, id: number) => {
   try {
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
     const response = await fetch(`http://localhost:3000/tasks/${id}`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
+        Authorization: `Bearer ${String(token)}`,
       },
       body: JSON.stringify(body),
     });
+
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
 
     const result = await response.json();
     console.log("Success:", result);
@@ -76,13 +108,24 @@ const updateTask = async (body: Body, id: number) => {
 
 const completedTask = async (id: number) => {
   try {
-    const response = await fetch(`http://localhost:3000/tasks/${id}/completed`, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
-      },
-    });
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
+    const response = await fetch(
+      `http://localhost:3000/tasks/${id}/completed`,
+      {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${String(token)}`,
+        },
+      }
+    );
+
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
 
     const result = await response.json();
     console.log("Success:", result);
@@ -93,13 +136,21 @@ const completedTask = async (id: number) => {
 
 const deleteTask = async (id: number) => {
   try {
+    const token = sessionStorage.getItem("token");
+    if (token === null) {
+      throw new Error("invalid jwt token");
+    }
     const response = await fetch(`http://localhost:3000/tasks/${id}/`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        "Authorization": `Bearer ${String(sessionStorage.getItem("token"))}`
+        Authorization: `Bearer ${String(token)}`,
       },
     });
+
+    if (!response.ok) {
+      throw new Error("Failed to fetch data");
+    }
 
     const result = await response.json();
     console.log("Success:", result);
@@ -109,10 +160,10 @@ const deleteTask = async (id: number) => {
 };
 
 export const TodoListService = {
-    getAllTask,
-    getTaskById,
-    createTask,
-    updateTask,
-    completedTask,
-    deleteTask
-}
+  getAllTask,
+  getTaskById,
+  createTask,
+  updateTask,
+  completedTask,
+  deleteTask,
+};
